implementa classe pedido e seus metodos do Crud e validações necessar…

+++ src/main/java/com/prata/venda/product/application/api/ProductListResponse.java
@@ -4,6 +4,7 @@ import com.prata.venda.product.domain.Product;
 import com.prata.venda.product.domain.Promotion;
 import lombok.Value;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -11,7 +12,7 @@ import java.util.stream.Collectors;
 public class ProductListResponse {
     Long idProduct;
     String name;
-    double price;
+    BigDecimal price;
     Promotion promotion;
 
     public static List<ProductListResponse> converte(List<Product> products){
